Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -15,7 +15,7 @@ int main() {
         if (b == 0) {
         printf("Error: Division by zero\n");
         } else {
-        printf("%.2f\n", (float)a / b);
+        printf("%.2f\n", a / b);
         }
     } else {
         printf("Error: Invalid operator\n");
